restructured unit testing
