+++ requirements.txt
@@ -18,7 +18,7 @@
 pandas==0.23.4
 scipy==1.3.1
 seaborn==0.9.0
-tsfresh
+tsfresh==0.12.0
 xlrd==1.2.0
 shapely
 matplotlib>=3.1.3
